Bouncy boys!

+++ animal3D-SDK/animal3D-SDK/source/animal3D-DemoPlugin/A3_DEMO/a3_Networking/Boid.cpp
@@ -2,7 +2,7 @@
 #include "A3_DEMO/a3_DemoState.h"
 #include "animal3D/animal3D.h"
 
-Boid::Boid(bool activeState, Vector2 newPos, Vector2 newVel, float newRot, int newRadius)
+Boid::Boid(bool activeState, Vector2 newPos, Vector2 newVel, float newRot, float newRadius)
 {
 	active = activeState;
 	position = newPos;
